Landing 0.9.57
Beginning to land the PACKAGING_20030906_BRANCH for bug 20640. Not part of the build, yet.
* * *
Continuing to land the PACKAGING_20030906_BRANCH for bug 20640. Not part of the build, yet.
* * *
Bug 231621 Add a middle-click action for the Chatzilla client window (open in new tab, configurable)
patch by Andreas Kunz (durbacher@gmx.de)
r=silver