Bug 1712002 - Prevent repeated handling of drop event. r=frg a=frg
When contacts are dragged from the address book and dropped into a
Mail&News compose window addressing field, the dropped content is
added to addressing fields via parseAndAddAddresses(), and is then
added a second time (to the drop target field) by the default drop
event handler.

Try to avoid this by calling .preventDefault() on the event when
DropOnAddressingTarget() is about to add dropped data.