Bug 1734407 - Prevent repeated handling of tab handle middle-click event. r=IanN a=IanN
When the browser is configured so that a middle-click will load a URL
from the clipboard or the primary selection, and such a URL is
present, middle-clicking a tab handle will close that tab and load the
URL in another tab.

contentAreaClick() (from suite/base/content/contentAreaClick.js),
which loads the URL, will check event.defaultPrevented before
proceeding, so use preventDefault() in the onTabClick() handler from
tabbrowser.xml.