Bug 1748178: Mail&News start.xhtml: point to the "Get Involved" page instead. r=IanN a=IanN
The "Welcome to SeaMonkey Mail&Newsgroups" page[1] that shows up in
the M&N main window message pane has a link (last paragraph, first
word, "We") that used to point to a mozilla.org webpage for the
mail-and-news client project[2].

The link target was changed to [3] in bug 1140353 as [2] was then
redirecting to a Thunderbird-specific page.

As now [3] doesn't work anymore, this changes the link target again,
this time to point to SeaMonkey's "Get Involved" web page.

[1] suite/mailnews/content/start.xhtml, viewable inside SeaMonkey as
    chrome://messenger/content/start.xhtml

[2] http://www.mozilla.org/mailnews/

[3] https://developer.mozilla.org/docs/Mozilla/Projects/MailNews