Bug 1752247 - Remove remaining rls arguments. r=darktrojan
In Firefox, the argument was removed from Google search params in
bug 1103216 and from other Yahoo Japan search params in bug 487407,
and seems to have stayed in the Yahoo Japan Auctions search params
as an oversight.

It seems the same happened in Thunderbird, and something similar in
Seamonkey for Google Japan.

Differential Revision: https://phabricator.services.mozilla.com/D137489