Bug 1328237 - Clear gSearchRemovedTerms after use and handle exceptions. r=IanN a=IanN
The code in searchTermOverlay.js handles removal of rules/rows in the
user interface by adding these to an array of removed entries
(gSearchRemovedTerms), which is used in saveSearchTerms() to update
the provided searchTerms.

This currently causes a problem in the message and address book search
dialogs, as the second run of saveSearchTerms() after a removal (from
the interface) will see indexOf() throw NS_ERROR_FAILURE because the
entry has been already removed from searchTerms.

To avoid the exception, with this change the contents of
gSearchRemovedTerms are discarded after being processed. It also adds
a try-catch around the problematic line, in case something else
arises, and replaces the use of dump() in the existing try-catch block
with Cu.reportError().